Small analysis binary invocation clean up
I don't like callbacks though
